Added 2 more maze types.
Because the mazes are generated by looping over the grid in a very straightforward column by column way, the mazes generated were biased to flow in the direction that the maze was generated. That is that they branched down and to the right.

To remedy this, I have added two alternative maze generation types which section up the space and then generate tiles.

The Convergent Quadrants method divides the graph into four quadrants and then generates tiles for each quadrant starting at the point farthest away from the center of the graph.

The Divergent Quadrants method divides the graph into four quadrants and then generates tiles for each quadrant starting at the point closest to the center of the graph.

The method used to generate the maze is randomly decided each time.